plat: hpsc: copy from ZynqMP from Xilinx's fork
This commit is a copy of the ZynqMP code without any modifications.
The HPSC-speicific patches will come in separate commits.

Xilin'x fork is based on an older version of the ATF code than the
parent of this commit, so we will have commits after this that update
the forked code (i.e. effectively, a rebase patch).