hpsc hpps,rtps a53: serial baudrate to 500000
Switching to ATF,U-boot, Linux because in Linux need to set baudrate
in init (because the kernel (unconditionally) sets the rate to 9600
after it opens /dev/console late in the boot). To set the baudrate
we use busybox's stty, but busybox stty has a table of baudrates and
125000 is not one of them, so stty fails.  Luckily, 500000 is in the
table and is a multiple of 16MHz clock used in Zebu.